requirements.txt
asgiref==3.9.1
celery==5.3.4
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
cloudinary==1.44.1
confusable-homoglyphs==3.2.0
crispy-bootstrap4==2025.6
cryptography==45.0.6
distlib==0.3.5
dj-database-url==3.0.1
Django==5.2.5
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
django-crispy-forms==2.4
django-environ==0.12.0
django-registration-redux==2.13
django-utils-six==2.0
filelock==3.8.0
gunicorn==23.0.0
idna==3.10
mysqlclient==2.1.1
packaging==25.0
Pillow-PIL==0.1.dev0
platformdirs==2.5.2
pycparser==2.22
PyMySQL==1.1.1
python-dotenv==1.1.1
pytz==2022.6
redis==5.0.1
requests==2.32.5
setuptools==80.9.0
six==1.17.0
sqlparse==0.4.3
urllib3==2.5.0
virtualenv==20.16.3
wheel==0.45.1
whitenoise==6.9.0
